+++ app/Services/Account/Activity/Activity/AttachContactToActivity.php
@@ -49,7 +49,7 @@ class AttachContactToActivity extends BaseService
      * @param array $data
      * @return Activity
      */
-    public function execute(array $data) : Activity
+    public function execute(array $data): Activity
     {
         $this->validate($data);