+++ app/Services/Contact/Contact/UpdateContactWork.php
@@ -29,7 +29,7 @@ class UpdateContactWork extends BaseService
      * @param array $data
      * @return Contact
      */
-    public function execute(array $data) : Contact
+    public function execute(array $data): Contact
     {
         $this->validate($data);