+++ app/Services/Contact/Occupation/DestroyOccupation.php
@@ -26,7 +26,7 @@ class DestroyOccupation extends BaseService
      * @param array $data
      * @return bool
      */
-    public function execute(array $data) : bool
+    public function execute(array $data): bool
     {
         $this->validate($data);