+++ routes/api.php
@@ -98,6 +98,7 @@ Route::group(['middleware' => ['auth:api']], function () {
         // Gifts
         Route::apiResource('gifts', 'ApiGiftController');
         Route::get('/contacts/{contact}/gifts', 'ApiGiftController@gifts');
+        Route::put('/gifts/{gift}/photo/{photo}', 'ApiGiftController@associate');
 
         // Debts
         Route::apiResource('debts', 'ApiDebtController');