+++ CHANGELOG.md
@@ -3,10 +3,11 @@
 ### New features:
 
 * Add ability to edit activities
+* Associate a photo to a gift
 
 ### Enhancements:
 
-*
+* Gift are now added and updated inline
 
 ### Fixes: